fix: disable dupImport check in gocritic for cgo false positives
Disable the gocritic dupImport check which incorrectly flags cgo
import "C" statements as duplicate imports. This is a known false
positive when using cgo - the "C" pseudo-package import is special
syntax and not a regular package import.

The linter was incorrectly reporting:
  dupImport: package is imported 2 times under different aliases

For files:
- dicom/pixel/jpeg2000.go
- dicom/pixel/jpeg_lossless.go

Each file only has a single valid `import "C"` statement for cgo.