fix: use libturbojpeg instead of libjpeg in CGO directive
Change the pkg-config directive in jpeg_lossless.go from 'libjpeg' to
'libturbojpeg'. This fixes build failures on Ubuntu where libturbojpeg
is the standard JPEG library.

libturbojpeg is a drop-in replacement for libjpeg that provides the
same jpeglib.h API and is available on both macOS and Ubuntu:
- macOS (Homebrew): jpeg-turbo package provides libturbojpeg
- Ubuntu: libturbojpeg0-dev package

Error was:
  Package libjpeg was not found in the pkg-config search path
  FAIL github.com/codeninja55/go-radx/benchmarks [build failed]

This fixes the test build failures in GitHub Actions.